Restructure InputComponent
